dev_requirements.txt
colorama~=0.4.4
django-extensions==3.1.5
django-filter==21.1
django==4
djangorestframework-simplejwt==4.8.0
djangorestframework==3.13.1
djoser==2.1.0
drf-extra-fields==3.2.1
gunicorn==20.1.0
Pillow==8.4.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytest==6.2.4
python-decouple==3.5
python-dotenv
pytz==2020.1
requests==2.26.0
sqlparse==0.3.1